Fix diff order to compare dst with src
The current diffs are confusing, since they compare the source against
the destination file. To make it less confusing: source is the file
that contains modifications, and destination is the one to be
overwritten.

With the current order, it displays changes as if you were applying
the destination file on the source file.

This commit makes it so that it displays the changes it would need to
apply source on destination. This is consistent with tools such as
patch, that take first the destination file, and second the patch
you're going to apply. Diffing should be done the same way, first the
destination, second the file with the new changes. This is how other
tools that provide diffs such as git also work.